+++ java-bootcamp/src/com/globant/javabootcamp/db/DBdummy.java
@@ -1,4 +1,4 @@
-package DB;
+package com.globant.javabootcamp.db;
 
 public class DBdummy {
 	
@@ -7,7 +7,7 @@ public class DBdummy {
 	public static boolean login(String user, String pass) {
 		
 		if (!connection)
-		if(user.equals("pepito")&(pass.equals("pepito"))) {
+		if(user.equals("SergioArrieta")&(pass.equals("abcd33"))) {
 			connection = true;
 			return true;
 		}
