+++ java-bootcamp/src/com/globant/javabootcamp/exercise2/Connection.java
@@ -1,4 +1,4 @@
-package exercise2_abstractFactory;
+package com.globant.javabootcamp.exercise2;
 
 public interface Connection {
