+++ java-bootcamp/src/com/globant/javabootcamp/exercise2/MicrosoftFactory.java
@@ -1,4 +1,4 @@
-package exercise2_abstractFactory;
+package com.globant.javabootcamp.exercise2;
 
 public class MicrosoftFactory implements AbstractFactory{
