+++ java-bootcamp/src/com/globant/javabootcamp/exercise2/MicrosoftODBC.java
@@ -1,4 +1,4 @@
-package exercise2_abstractFactory;
+package com.globant.javabootcamp.exercise2;
 
 public class MicrosoftODBC implements MicrosoftSQLServerConnection{
