+++ java-bootcamp/src/com/globant/javabootcamp/exercise2/MicrosoftSQLServerConnection.java
@@ -1,4 +1,4 @@
-package exercise2_abstractFactory;
+package com.globant.javabootcamp.exercise2;
 
 public interface MicrosoftSQLServerConnection extends Connection  {
