+++ java-bootcamp/src/com/globant/javabootcamp/exercise2/OracleConnection.java
@@ -1,4 +1,4 @@
-package exercise2_abstractFactory;
+package com.globant.javabootcamp.exercise2;
 
 public interface OracleConnection extends Connection{
