+++ java-bootcamp/src/com/globant/javabootcamp/exercise2/OracleODBC.java
@@ -1,4 +1,4 @@
-package exercise2_abstractFactory;
+package com.globant.javabootcamp.exercise2;
 
 public class OracleODBC implements OracleConnection{
