+++ java-bootcamp/src/com/globant/javabootcamp/exercise3/Client.java
@@ -1,4 +1,4 @@
-package exercise3_proxy;
+package com.globant.javabootcamp.exercise3;
 
 public class Client {
