+++ java-bootcamp/src/com/globant/javabootcamp/exercise3/RealDBConnection.java
@@ -1,6 +1,6 @@
-package exercise3_proxy;
+package com.globant.javabootcamp.exercise3;
 
-import DB.DBdummy;
+import com.globant.javabootcamp.db.DBdummy;
 
 public class RealDBConnection implements DBInterface{
