+++ java-bootcamp/src/com/globant/javabootcamp/exercise4/MicrosoftSQLConnectionBuilder.java
@@ -1,4 +1,4 @@
-package exercise4_builder;
+package com.globant.javabootcamp.exercise4;
 
 public class MicrosoftSQLConnectionBuilder extends AbstractBuilder {
 	//Concret Builder
